refactor: reorganize AI components following Magento best practices - Move AI/Api/* to Api/AI/* (API interfaces) - Move AI/Service/* to Service/AI/* (business logic) - Move AI/Model/* to Model/* (data models) - Move AI/Controller/* to Controller/AI/* (controllers) - Update all namespaces and imports accordingly
