+++ Model/ChatMessage.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace Bold\CheckoutPaymentBooster\AI\Model;
+namespace Bold\CheckoutPaymentBooster\Model;
 
-use Bold\CheckoutPaymentBooster\AI\Api\Data\ChatMessageInterface;
+use Bold\CheckoutPaymentBooster\Api\AI\Data\ChatMessageInterface;
 use Magento\Framework\DataObject;
 
 /**
