+++ etc/webapi.xml
@@ -67,19 +67,19 @@
     
     <!-- AI Agent Endpoints -->
     <route url="/V1/ai-agent/chat/send" method="POST">
-        <service class="Bold\CheckoutPaymentBooster\AI\Api\ChatInterface" method="sendMessage"/>
+        <service class="Bold\CheckoutPaymentBooster\Api\AI\ChatInterface" method="sendMessage"/>
         <resources>
             <resource ref="anonymous"/>
         </resources>
     </route>
     <route url="/V1/ai-agent/chat/history/:sessionId" method="GET">
-        <service class="Bold\CheckoutPaymentBooster\AI\Api\ChatInterface" method="getChatHistory"/>
+        <service class="Bold\CheckoutPaymentBooster\Api\AI\ChatInterface" method="getChatHistory"/>
         <resources>
             <resource ref="anonymous"/>
         </resources>
     </route>
     <route url="/V1/ai-agent/chat/start" method="POST">
-        <service class="Bold\CheckoutPaymentBooster\AI\Api\ChatInterface" method="startSession"/>
+        <service class="Bold\CheckoutPaymentBooster\Api\AI\ChatInterface" method="startSession"/>
         <resources>
             <resource ref="anonymous"/>
         </resources>
